requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.2
httpx>=0.27.0
websockets>=12.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
minio>=7.2.7
sse-starlette>=2.1.2
python-multipart>=0.0.9
orjson>=3.10.0
